ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.6
langchain-community==0.0.19
chromadb==0.4.22
pypdf==4.0.1
python-multipart==0.0.9
httpx==0.26.0
pydantic==2.6.0
tiktoken==0.6.0
